add plan.py, time_window.py and utils.py to greedy_max module so __ini__.py is decluttered
